built header with its design and revised style
The header section has been built along with its style.

There has been a mistake in style which caused grammatical mistakes. It was text-transformed, so I have removed it.

Some random generic background is and will be added to each section and will be removed upon completion of the project. This is for visionary clue of the design.