categories section built and nav scroll fixed
The categories section has been built along with its design so far. Earlier  I wrote the js code to change the background color of nav on scroll but it didn't return to its initial background color once the scroll was back at the top. A bit of java script fixed it.