faqs has been completed
Frequently asked questions sections has been completed with its design.

js code has been written to expand the answer, but I couldn't add much of smooth animation to show the answer.

besides, changing a css property (box-sizing) to border-box brought an issue with the layout. So it needs to be fixed in next commit.