Responsive nav bar, js for showing and hiding nav
The navigation menu has been made responsive.
The js code to make the menu button show and hide has been written.
The animation has been added to show the menu. The animation base is rotatez, rotatex and scale.
For the element to stay in the last point of animation we should set the animation direction to forward.