Minor changes on nav, footer and team social links
The nav links gets a border bottom on hover.
The footer links color was to sharp eyes. So it has been changed from dark pinkish color to the white shady color.
The vertical positioning of team__member social links has set using a bit of complicated styling. So I just used the translateY(-50%).