contact.html made fully response
The container div's width has been set to relative units which helps in being a bit of responsive.

The container has been made responsive in two different breakpoints which are 1262px and 800px.

In case the screen hits 1262px, the firstname and lastname input width get full width of form.

And in case of 800px the form container grid columns changes to a single one.