set up FormSpree to actually send/receive email
The FormSpree url added to form action and method is set to post. This way, the email will actually be sent and received on the other end.

A minor problem was there that used to capitalize every word of the message and first letter of email, so fixed is using text transform.