Created courses.html with its responsive design
The courses.html file has been built. It has categories button which is clickable and take you to its relevant category courses. The data, however is the same and random in each category, but it works well.

The navigation menu was showing up even on window scroll, so added javascript code on index.js to hide the menu on scroll.