built sample_course, and its design/no responsive
The sample_course.html files has been built. The css design has been applied. It is not responsive, however.

This page is to show any course which is clicked on to show its complete information about the course. It also contains an enroll button. This button will then enroll the user to the corresponding course then.