Added login form to all pages+design+js code
Added the necessary codes for login form.
Wrote the CSS design code for login form, and an overlay div which gets the blur effect.
Wrote the JS code to bring the login form out on the window, and vanish back when the cross button is clicked.

The notebook button on site's logo has been changed to the book icons.