Added the dark/light mode button and customization
Wrote the Markup for dark/light mode button, its design and js code.

The CSS files needed a bit of customization for dark mode, so customized the other pages CSS files a bit.

The forms required inputs were given the 'required' attribute.

The featured course image was losing its quality while in small devices, so brought a bit of changes in CSS files to shrink in small devices.