stable dark/light mode, selected links,legit image
The first thing I have done is that I added legit image to the courses.html page.  The images belongs to the categories and topics of the course.

The second thing is adding selected class to nav links and category buttons in courses.html page. The code adds a bit of extra effect on selected category button or link. The corresponding code has been added to index.js and courses.js files.

The last thing I have done is all about how to keep the light or either dark mode even in you traverse through different pages of the website. I used the local storage to keep track of whether the website should be light mode or dark mode.